+++ Makefile
@@ -11,7 +11,7 @@ bench:
 
 .PHONY: protoc
 protoc:
-	protoc --go_out=module=github.com/openzipkin/zipkin-go:. proto/v2/zipkin.proto
+	protoc --go_out=module=github.com/openzipkin/zipkin-go:. proto/zipkin_proto3/zipkin.proto
 	protoc --go_out=module=github.com/openzipkin/zipkin-go:. proto/testing/service.proto
 	protoc --go-grpc_out=module=github.com/openzipkin/zipkin-go:. proto/testing/service.proto
