+++ proto/zipkin_proto3/encode_proto.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v2
+package zipkin_proto3
 
 import (
 	"encoding/binary"
