+++ proto/zipkin_proto3/encode_proto_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v2_test
+package zipkin_proto3_test
 
 import (
 	"encoding/json"
@@ -22,7 +22,7 @@ import (
 	"time"
 
 	zipkinmodel "github.com/openzipkin/zipkin-go/model"
-	zipkin_proto3 "github.com/openzipkin/zipkin-go/proto/v2"
+	zipkin_proto3 "github.com/openzipkin/zipkin-go/proto/zipkin_proto3"
 )
 
 func TestExportSpans(t *testing.T) {
